the getPostLikes was in the wrong controller
