+++ controllers/profileScreenController.js
@@ -59,27 +59,3 @@ export const getUserPosts = async (req, res) => {
     res.status(error.statusCode).json(error.message);
   }
 };
-
-export const getPostLikes = async (req, res) => {
-  try {
-    const post = await postRating.findById(req.query.post_id);
-    const likes = await getAllUserLikes(post.likes, req);
-    res.status(200).json(likes);
-  } catch (error) {
-    res.status(error.statusCode).json(error.message);
-  }
-};
-
-const getAllUserLikes = (userLikes, req) => {
-  const accessToken = getAccessToken(req);
-  const spotifyApi = new SpotifyWebApi();
-  spotifyApi.setAccessToken(accessToken);
-  return Promise.all(
-    userLikes.map(
-      async (user_id) =>
-        await spotifyApi
-          .getUser(user_id)
-          .then((data) => ({ id: data.body.id, display_name: data.body.display_name, image_url: data.body.images[0]?.url }))
-    )
-  );
-};
