+++ routes/albumDetailsRoutes.js
@@ -12,5 +12,6 @@ router.get("/getUsersProfile", albumDetailsController.getUsersProfile);
 router.post("/createPost", albumDetailsController.createPost);
 router.delete("/deletePost", albumDetailsController.deletePost);
 router.patch("/handleLikes", albumDetailsController.handleLikes);
+router.get("/getPostLikes", albumDetailsController.getPostLikes);
 
 export default router;
